ci: test linux/arm64 natively and guard test-code portability
The first run of the new lane failed: internal/sandbox's test package could not compile on linux/arm64 at all, because coverage_deep_test.go was tagged plain linux while the seccomp filter builder and BPF helpers it calls exist only under linux and amd64. A published target had zero sandbox test coverage and nothing reported it, because go build never compiles test files. The seccomp-dependent tests move to a linux and amd64 file, and the lint job now cross-targets go vet, which does compile test files, so the class is caught for every published non-Windows target from an amd64 runner. Windows is deliberately excluded pending its own two instances.
