+++ .github/workflows/ci.yaml
@@ -637,6 +637,33 @@ jobs:
       - name: golangci-lint (enterprise)
         run: golangci-lint run --build-tags enterprise ./...
 
+      # Test files are compiled by `go test`, never by `go build`, so a test file
+      # that references a symbol guarded to one architecture makes an entire test
+      # package uncompilable there while every build job stays green. That is not
+      # hypothetical: internal/sandbox's test package could not compile on
+      # linux/arm64 at all, because coverage_deep_test.go was tagged plain linux
+      # while the seccomp filter builder and BPF helpers it calls exist only under
+      # linux && amd64. A published target therefore had zero sandbox test
+      # coverage and nothing said so. The same shape hit internal/config on
+      # Windows previously via syscall.Mkfifo.
+      #
+      # `go vet` compiles test files, so cross-targeting it catches the class on
+      # this amd64 runner without needing hardware for every target. It runs here
+      # rather than in each native lane so a target with no lane is still covered.
+      #
+      # windows/amd64 and windows/arm64 are deliberately NOT in this list yet:
+      # internal/cli/contain and internal/emit have test files with this exact
+      # defect on Windows today, tracked separately. Adding them here before that
+      # is fixed would land a knowingly-red required check, and a check people
+      # expect to be red stops being read. Add both the moment that lands.
+      - name: Test code compiles for every published non-Windows target
+        run: |
+          set -euo pipefail
+          for target in linux/arm64 darwin/amd64 darwin/arm64; do
+            printf '==> %s\n' "$target"
+            GOOS="${target%%/*}" GOARCH="${target##*/}" go vet ./...
+          done
+
       - name: Python pin policy
         run: bash scripts/check-python-pins.sh
